refactor: extract answer paragraph parsing into parseAnswer.js
AnswerPage (next commit) needs the same paragraph/marker parsing to know
which citations belong beside which paragraph for margin notes; pulling
it out of AnswerPassage avoids duplicating the marker-splitting logic.
